backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
pydantic==2.5.0
httpx==0.25.1

# New for Phase 2
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Optional for future
# openai>=1.3.0
# anthropic>=0.7.0
